+++ src/components/dom-overlays/ArMarkerOverlay.svelte
@@ -1,0 +1,12 @@
+<!--
+  (c) 2021 Open AR Cloud
+  This code is licensed under MIT license (see LICENSE for details)
+-->
+
+<!-- DOM-overlay on top of AR canvas AR mode is marker -->
+<script>
+    import { pointMarkerMessage } from '@src/contentStore';
+</script>
+
+
+<p>{$pointMarkerMessage}</p>
